+++ tests/python/converter_test.py
@@ -3,7 +3,7 @@
 
 Unit tests of the converter class. This module uses the common test specifications defined in
 psdi_data_conversion/testing/conversion_test_specs.py so that a common set of conversion tests is performed through
-the Python library (this module), the command-line application, and the GUI.
+the Python library (this module), the command-line interface, and the GUI.
 """
 
 import logging