+++ CHANGELOG.md
@@ -200,7 +200,7 @@
 
 ### New and Changed Functionality
 
-- When listing formats supported by a given converter in the command-line application, the description of each format will also be shown in the table
+- When listing formats supported by a given converter in the command-line interface, the description of each format will also be shown in the table
 - A warning will now be printed to stderr and logged if an unrecognised format flag or option is provided for conversion with Open Babel
 
 ### Bugfixes